% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals_by.R
\name{individuals_by}
\alias{individuals_by}
\title{"individuals_by"}
\usage{
individuals_by(
  individual_data = individuals,
  ind_id = nom,
  group = NULL,
  group_other = NULL,
  group_other_name = NULL,
  font_size = 7,
  height = NULL,
  width = NULL,
  colour_scale = JS("d3.scaleOrdinal(d3.schemeCategory20);"),
  font_family = "serif",
  link_distance = 50,
  link_width = JS("function(d) { return Math.sqrt(d.value); }"),
  radius_calculation = "4*Math.sqrt(d.nodesize)+2",
  charge = -30,
  link_colour = "#666",
  opacity = 0.6,
  zoom = FALSE,
  arrows = FALSE,
  bounded = FALSE,
  display_labels = 0,
  click_action = NULL
)
}
\arguments{
\item{individual_data}{The data frame containing the individual-level data. By default, it assumes a data frame called \code{individuals}.}

\item{ind_id}{The variable/column in the \code{individual_data} data frame that represents the individual identifier. By default, it assumes a column named "nom".}

\item{group}{The variable/column in the \code{individual_data} data frame that represents the grouping variable. It is used to assign colors to the nodes in the visualization. If not provided, the visualization will not group the nodes.}

\item{group_other}{is a string representing the name of the column in the "individual_data" data block that contains additional group information for each individual.}

\item{group_other_name}{is a string representing a custom name for the "group_other" column to display in the plot.}

\item{font_size}{numeric font size in pixels for the node text labels.By default, it is set to 7.}

\item{height}{numeric height for the network graph's frame area in pixels.}

\item{width}{numeric width for the network graph's frame area in pixels.}

\item{colour_scale}{character string specifying the categorical colour scale for the nodes. See \code{https://github.com/d3/d3/blob/master/API.md#ordinal-scales}.}

\item{font_family}{font family for the node text labels.}

\item{link_distance}{numeric or character string. Either numberic fixed distance between the links in pixels (actually arbitrary relative to the diagram's size). Or a JavaScript function, possibly to weight by Value. For example: linkDistance = JS("function(d){return d.value * 10}").}

\item{link_width}{numeric or character string. Can be a numeric fixed width in pixels (arbitrary relative to the diagram's size). Or a JavaScript function, possibly to weight by Value. The default is linkWidth = JS("function(d) { return Math.sqrt(d.value); }").}

\item{radius_calculation}{character string. A javascript mathematical expression, to weight the radius by Nodesize. The default value is radiusCalculation = JS("Math.sqrt(d.nodesize)+6").}

\item{charge}{numeric value indicating either the strength of the node repulsion (negative value) or attraction (positive value).}

\item{link_colour}{character vector specifying the colour(s) you want the link lines to be. Multiple formats supported (e.g. hexadecimal).}

\item{opacity}{numeric value of the proportion opaque you would like the graph elements to be.}

\item{zoom}{logical value to enable (TRUE) or disable (FALSE) zooming.}

\item{arrows}{logical value to enable directional link arrows.}

\item{bounded}{logical value to enable (TRUE) or disable (FALSE) the bounding box limiting the graph's extent. See \code{http://bl.ocks.org/mbostock/1129492}.}

\item{display_labels}{is a numeric value representing the number of characters of the label to display on each node.}

\item{click_action}{character string with a JavaScript expression to evaluate when a node is clicked.}
}
\value{
Returns a network graph object
}
\description{
individuals_by function creates a force-directed network visualization using the forceNetwork function from the networkD3 package.
}
\examples{
# TODO
}
