% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiative_by.R
\name{initiative_by}
\alias{initiative_by}
\title{"Initiative_by"}
\usage{
initiative_by(
  initiative_data,
  by = "pays",
  filter_var = NULL,
  filter_vals = NULL,
  node_size = c("type", "age"),
  group = NULL,
  font_size = 7,
  height = NULL,
  width = NULL,
  colour_scale = JS("d3.scaleOrdinal(d3.schemeCategory20);"),
  font_family = "serif",
  link_distance = 50,
  link_width = JS("function(d) { return Math.sqrt(d.value); }"),
  radius_calculation = "4*Math.sqrt(d.nodesize)+2",
  charge = -30,
  link_colour = "#666",
  opacity = 0.6,
  zoom = FALSE,
  arrows = FALSE,
  bounded = FALSE,
  display_labels = 0,
  click_action = NULL
)
}
\arguments{
\item{initiative_data}{a data frame object containing the initiative data}

\item{by}{}

\item{filter_var}{}

\item{filter_vals}{}

\item{node_size}{character string specifying the a column in the \code{initiative_data} data frame with some value to vary the node radius's with. See also \code{radiusCalculation}.}

\item{group}{character string specifying the group of each node in the \code{initiative_data} data frame.}

\item{font_size}{}

\item{height}{}

\item{width}{}

\item{colour_scale}{}

\item{font_family}{}

\item{link_distance}{}

\item{link_width}{}

\item{radius_calculation}{}

\item{charge}{}

\item{link_colour}{}

\item{opacity}{}

\item{zoom}{}

\item{arrows}{}

\item{bounded}{}

\item{display_labels}{}

\item{click_action}{}
}
\value{
Returns a network graph object
}
\description{
This function is a wrapper function for \code{networkD3::forceNetwork}.
explain what it does -
}
\examples{
# todo
}
