% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_data}
\alias{map_data}
\title{Map Data for Network Visualisation}
\usage{
map_data(data, values, node_values)
}
\arguments{
\item{data}{The original data frame.}

\item{values}{The column name representing the primary values to be mapped.}

\item{node_values}{A character vector specifying the column names representing the node values.}
}
\value{
A network visualisation generated using the \code{forceNetwork} function.
}
\description{
This function takes a data frame and generates a network visualization using the \code{forceNetwork} function
from the \code{networkD3} package. It separates and expands the data based on the specified values and node values,
and creates links between them.
}
\examples{
# An example with simple data 
df_original <- data.frame(initiative = c("A", "B", "C", "D"),
                          country = c("BF", "BF Niger", "Niger", "Niger"))
map_data(df_original, "initiative", "country")


df_original2 <- data.frame(individuals = c("A", "B", "C", "D"),
                           institutions = c("a b", "a", "a", "b c"),
                           initiatives = c("z", "z", "y z", "y"))
map_data(df_original2, "individuals", c("institutions", "initiatives"))

}
