% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_other.R
\name{replace_other}
\alias{replace_other}
\title{"replace_other"}
\usage{
replace_other(
  data = individuals,
  group = NULL,
  group_other = paste0("autre_", group),
  group_other_name = NULL,
  RAS = FALSE
)
}
\arguments{
\item{data}{The data frame in which the replacement will be performed. By default, it assumes a data frame called \code{individuals}.}

\item{group}{The name of the group variable to be replaced. This parameter is required.}

\item{group_other}{The name of the replacement group variable. It is constructed by appending the prefix "autre_" to the original group name. For example, if \code{group} is "pays", then \code{group_other} becomes "autre_pays".}

\item{group_other_name}{The specific value within the \code{group} variable that should be replaced. When this value is encountered in the \code{group} variable, it will be replaced with the corresponding value from \code{group_other}. This parameter is required.}

\item{RAS}{A logical value indicating whether the replacement should be performed with "RAS" when the \code{group_other_name} is encountered. If \code{RAS} is set to \code{TRUE}, the replacement will be "RAS"; otherwise, it will be the corresponding value from \code{group_other}. By default, it is set to \code{FALSE}.}
}
\value{
Returns data
}
\description{
This function replaces a specific group in a data frame with another group.
}
\note{
The \code{mutate} function is used to create a new column in the data frame named \code{group_var}, which replaces the values based on the provided conditions. When \code{RAS} is set to \code{TRUE}, it checks if the value in the \code{group} variable is equal to \code{group_other_name}. If it is, the corresponding value from \code{group_other} is assigned to \code{group_var}. Additionally, if the value in the \code{group} variable is \code{NA} (missing), it is replaced with "RAS". If \code{RAS} is set to \code{FALSE}, the replacement is performed without using "RAS".
}
\examples{
# TODO
}
